Port e2e tests to UI integration tests
